@@ -1,9 +1,8 @@
-﻿namespace Krompaco.RecordCollector.Content.FrontMatterParsers
+﻿namespace Krompaco.RecordCollector.Content.FrontMatterParsers;
+
+public class ParserBase
 {
-    public class ParserBase
-    {
 #pragma warning disable CS8618 // Non-nullable field must contain a non-null value when exiting constructor. Consider declaring as nullable.
-        protected string FullName { get; set; }
+    protected string FullName { get; set; }
 #pragma warning restore CS8618 // Non-nullable field must contain a non-null value when exiting constructor. Consider declaring as nullable.
-    }
 }
