@@ -1,59 +1,58 @@
 ﻿using Krompaco.RecordCollector.Content.Models;
 
-namespace Krompaco.RecordCollector.Content.FrontMatterParsers
+namespace Krompaco.RecordCollector.Content.FrontMatterParsers;
+
+public class TypeDetector
 {
-    public class TypeDetector
+    private readonly StreamReader tr;
+
+    public TypeDetector(StreamReader tr)
     {
-        private readonly StreamReader tr;
+        this.tr = tr;
+    }
 
-        public TypeDetector(StreamReader tr)
+    public FrontMatterType GetFrontMatterType()
+    {
+        var rowCount = 0;
+
+        while (this.tr.Peek() >= 0)
         {
-            this.tr = tr;
-        }
+            rowCount++;
+            var line = this.tr.ReadLine() ?? string.Empty;
 
-        public FrontMatterType GetFrontMatterType()
-        {
-            var rowCount = 0;
-
-            while (this.tr.Peek() >= 0)
+            if (line.StartsWith("{", StringComparison.Ordinal))
             {
-                rowCount++;
-                var line = this.tr.ReadLine() ?? string.Empty;
-
-                if (line.StartsWith("{", StringComparison.Ordinal))
-                {
-                    return this.ResetStreamAndReturn(FrontMatterType.Json);
-                }
-
-                if (line.Trim() == "+++")
-                {
-                    return this.ResetStreamAndReturn(FrontMatterType.Toml);
-                }
-
-                if (line.Trim() == "---")
-                {
-                    return this.ResetStreamAndReturn(FrontMatterType.Yaml);
-                }
-
-                if (line.TrimStart().StartsWith("<html", StringComparison.OrdinalIgnoreCase))
-                {
-                    return this.ResetStreamAndReturn(FrontMatterType.HtmlDocument);
-                }
-
-                if (rowCount == 5)
-                {
-                    break;
-                }
+                return this.ResetStreamAndReturn(FrontMatterType.Json);
             }
 
-            return this.ResetStreamAndReturn(FrontMatterType.MarkdownDocument);
+            if (line.Trim() == "+++")
+            {
+                return this.ResetStreamAndReturn(FrontMatterType.Toml);
+            }
+
+            if (line.Trim() == "---")
+            {
+                return this.ResetStreamAndReturn(FrontMatterType.Yaml);
+            }
+
+            if (line.TrimStart().StartsWith("<html", StringComparison.OrdinalIgnoreCase))
+            {
+                return this.ResetStreamAndReturn(FrontMatterType.HtmlDocument);
+            }
+
+            if (rowCount == 5)
+            {
+                break;
+            }
         }
 
-        private FrontMatterType ResetStreamAndReturn(FrontMatterType fmt)
-        {
-            this.tr.BaseStream.Position = 0;
-            this.tr.DiscardBufferedData();
-            return fmt;
-        }
+        return this.ResetStreamAndReturn(FrontMatterType.MarkdownDocument);
+    }
+
+    private FrontMatterType ResetStreamAndReturn(FrontMatterType fmt)
+    {
+        this.tr.BaseStream.Position = 0;
+        this.tr.DiscardBufferedData();
+        return fmt;
     }
 }
