@@ -1,15 +1,14 @@
 ﻿using System.Globalization;
 
-namespace Krompaco.RecordCollector.Content.Languages
+namespace Krompaco.RecordCollector.Content.Languages;
+
+public class ContentCultureService
 {
-    public class ContentCultureService
+    #pragma warning disable CA1822 // Mark members as static
+    public bool DoesCultureExist(string cultureName)
     {
-        #pragma warning disable CA1822 // Mark members as static
-        public bool DoesCultureExist(string cultureName)
-        {
-            return CultureInfo.GetCultures(CultureTypes.AllCultures)
-                .Any(culture => string.Equals(culture.Name, cultureName, StringComparison.CurrentCultureIgnoreCase));
-        }
-        #pragma warning restore CA1822 // Mark members as static
+        return CultureInfo.GetCultures(CultureTypes.AllCultures)
+            .Any(culture => string.Equals(culture.Name, cultureName, StringComparison.CurrentCultureIgnoreCase));
     }
+    #pragma warning restore CA1822 // Mark members as static
 }
