@@ -1,15 +1,14 @@
-﻿namespace Krompaco.RecordCollector.Content.Models
-{
-    public enum ContentType
-    {
-        /// <summary>
-        /// HTML content.
-        /// </summary>
-        Html,
+﻿namespace Krompaco.RecordCollector.Content.Models;
 
-        /// <summary>
-        /// Markdown content.
-        /// </summary>
-        Markdown,
-    }
+public enum ContentType
+{
+    /// <summary>
+    /// HTML content.
+    /// </summary>
+    Html,
+
+    /// <summary>
+    /// Markdown content.
+    /// </summary>
+    Markdown,
 }
