@@ -1,30 +1,29 @@
-﻿namespace Krompaco.RecordCollector.Content.Models
+﻿namespace Krompaco.RecordCollector.Content.Models;
+
+public enum FrontMatterType
 {
-    public enum FrontMatterType
-    {
-        /// <summary>
-        /// JSON.
-        /// </summary>
-        Json,
+    /// <summary>
+    /// JSON.
+    /// </summary>
+    Json,
 
-        /// <summary>
-        /// TOML.
-        /// </summary>
-        Toml,
+    /// <summary>
+    /// TOML.
+    /// </summary>
+    Toml,
 
-        /// <summary>
-        /// YAML.
-        /// </summary>
-        Yaml,
+    /// <summary>
+    /// YAML.
+    /// </summary>
+    Yaml,
 
-        /// <summary>
-        /// No front matter found, complete HTML document.
-        /// </summary>
-        HtmlDocument,
+    /// <summary>
+    /// No front matter found, complete HTML document.
+    /// </summary>
+    HtmlDocument,
 
-        /// <summary>
-        /// No front matter found, complete Markdown document.
-        /// </summary>
-        MarkdownDocument,
-    }
+    /// <summary>
+    /// No front matter found, complete Markdown document.
+    /// </summary>
+    MarkdownDocument,
 }
