@@ -1,10 +1,9 @@
 ﻿using Krompaco.RecordCollector.Content.Models;
 
-namespace Krompaco.RecordCollector.Web.Models
+namespace Krompaco.RecordCollector.Web.Models;
+
+public interface IHasCurrentPage<TModel>
+    where TModel : SinglePage
 {
-    public interface IHasCurrentPage<TModel>
-        where TModel : SinglePage
-    {
-        TModel? CurrentPage { get; set; }
-    }
+    TModel? CurrentPage { get; set; }
 }
