@@ -1,7 +1,6 @@
-﻿namespace Krompaco.RecordCollector.Web.Resources
+﻿namespace Krompaco.RecordCollector.Web.Resources;
+
+// Dummy class to group shared resources
+public class SharedResource
 {
-    // Dummy class to group shared resources
-    public class SharedResource
-    {
-    }
 }
